# OpenApi.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Car Management API",
    "description": "Open API Documentation",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/register": {
      "post": {
        "summary": "Registrasi Member",
        "description": "Registrasi Member",
        "tags": [
          "Auth"
        ],
        "operationId": "Registrasi",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/RegistrasiRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Register Successfully",
            "headers": {},
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Created"
                }
              }
            }
          },
          "400": {
            "description": "Register Gagal",
            "headers": {},
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Register Gagal"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "definitions": {
    "RegistrasiRequest": {
      "title": "RegistrasiRequest",
      "example": {
        "email": "khairil@gmail.com",
        "password": "khairil123"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
...
  },
  "security": [],
...
}